% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBIBD1.R
\name{PBIBD1}
\alias{PBIBD1}
\title{Affine resolvable PBIB designs}
\usage{
PBIBD1(v, p, q)
}
\arguments{
\item{v}{Total number of treatments}

\item{p}{Positive integer (>=2)}

\item{q}{Positive integer (>=2)}
}
\value{
This function generates a PBIB design and its parameters, variance factors and efficiency factor.
}
\description{
This function generates a new series of affine resolvable Partially Balanced Incomplete Block Designs (PBIBDs) and its parameters (v, b1, b2, r, k1, k2), canonical efficiency factors, variance factor between associates and average variance factors.
}
\examples{
library(ResPBIBD)
PBIBD1(12, 4, 3)
}
