% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBIBD3.R
\name{PBIBD3}
\alias{PBIBD3}
\title{Resolvable PBIB designs}
\usage{
PBIBD3(v, p)
}
\arguments{
\item{v}{Total number of treatments}

\item{p}{Positive integer (>=2)}
}
\value{
This function generates a PBIB design and its parameters, variance factors and efficiency factor.
}
\description{
This function generates a new series of resolvable Partially Balanced Incomplete Block Designs (PBIBDs) and its parameters (v, b1, b2, r, k1, k2), canonical efficiency factors, variance factor between associates and average variance factors.
}
\examples{
library(ResPBIBD)
PBIBD3(12, 4)
}
